requirement.txt
Package                Version
---------------------- --------
alembic                1.9.4
aniso8601              9.0.1
autopep8               2.0.2
click                  8.1.3
configparser           5.3.0
Flask                  2.2.3
Flask-Migrate          4.0.4
Flask-RESTful          0.3.9
Flask-SQLAlchemy       3.0.3
greenlet               2.0.2
itsdangerous           2.1.2
Jinja2                 3.1.2
Mako                   1.2.4
MarkupSafe             2.1.2
mysql-connector-python 8.0.32
mysqlclient            2.1.1
pip                    22.0.2
protobuf               3.20.3
pycodestyle            2.10.0
python-dotenv          1.0.0
pytz                   2022.7.1
setuptools             59.6.0
six                    1.16.0
SQLAlchemy             2.0.4
tomli                  2.0.1
typing_extensions      4.5.0
Werkzeug               2.2.3
